projects/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "httpx>=0.28.1",
    "pgqueuer>=0.25.3",
    "pinecone>=8.0.0",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[google]>=1.44.0",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "reportlab>=4.4.9",
    "ruff>=0.14.12",
    "testcontainers[postgres]>=4.14.0",
]
